Have a won and lose.
Conditionally the indoors scene is missing my canvas ui for the heart containers and the player is the the most updated prefab so its missing some scripts. i still have a few fixes to make. the blue apple causes an automatic loss for the player from items. the main menu now has a proper credits and a controls list. more to be added before my final presentation. any consults for sound just ask :)