Banks Update
Final update for the night, this is very playable in my opinion. Music is functioning, menu dings, everything from before is re implemented. Obviously footsteps need to change when on different surfaces, but I do not have the knowledge or time to do so at this moment. My coding skill has already improved 10x.